Update Application class + nuke Watch command
