+++ src/allejo/stakx/Command/BuildCommand.php
@@ -7,19 +7,48 @@
 
 namespace allejo\stakx\Command;
 
-use allejo\stakx\Exception\FileAwareException;
-use allejo\stakx\Utilities\StrUtils;
+use allejo\stakx\Configuration;
+use allejo\stakx\Filesystem\File;
+use allejo\stakx\Filesystem\FilesystemLoader as fs;
+use allejo\stakx\Service;
+use allejo\stakx\Website;
 use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
-class BuildCommand extends BuildableCommand
+/**
+ * Class BuildableCommand.
+ *
+ * This abstract class handles configuring the website object
+ */
+abstract class BuildableCommand extends ContainerAwareCommand
 {
+    const NO_CONF = 'no-conf';
+    const NO_CLEAN = 'no-clean';
+    const USE_DRAFTS = 'use-drafts';
+    const WATCHING = 'watching';
+    const USE_CACHE = 'use-cache';
+    const SAFE_MODE = 'safe';
+    const BUILD_PROFILE = 'profile';
+
+    /** @var Configuration */
+    protected $configuration;
+
+    /** @var Website */
+    protected $website;
+
     /**
      * {@inheritdoc}
      */
     protected function configure()
     {
-        parent::configure();
+        $this->addOption('conf', 'c', InputOption::VALUE_REQUIRED, 'The configuration file to be used', Configuration::DEFAULT_NAME);
+        $this->addOption(self::SAFE_MODE, 's', InputOption::VALUE_NONE, 'Disable file system access from Twig');
+        $this->addOption(self::NO_CONF, 'l', InputOption::VALUE_NONE, 'Build a stakx website without a configuration file');
+        $this->addOption(self::NO_CLEAN, 'x', InputOption::VALUE_NONE, "Don't clean the _site before recompiling the website");
+        $this->addOption(self::USE_DRAFTS, 'd', InputOption::VALUE_NONE, 'Publish all ContentItems marked as drafts');
+        $this->addOption(self::USE_CACHE, null, InputOption::VALUE_NONE, 'Use the existing cache folder for building the website');
+        $this->addOption(self::BUILD_PROFILE, null, InputOption::VALUE_NONE, 'Display a profile of Twig templates');
 
         $this->setName('build');
         $this->setDescription('Builds the stakx website');
@@ -30,7 +59,20 @@ class BuildCommand extends BuildableCommand
      */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
-        parent::execute($input, $output);
+        $this->website = new Website($this->getContainer());
+        $this->website->setConfLess($input->getOption(self::NO_CONF));
+
+        $flags = [
+            self::SAFE_MODE, self::NO_CONF, self::NO_CLEAN,
+            self::USE_DRAFTS, self::USE_CACHE, self::BUILD_PROFILE,
+        ];
+
+        foreach ($flags as $flag)
+        {
+            $this->setServiceParameter($input, $flag);
+        }
+
+        $this->configureConfigurationFile($input);
 
         try
         {
@@ -60,5 +102,38 @@ class BuildCommand extends BuildableCommand
         }
 
         return 1;
+    }
+
+    private function configureConfigurationFile(InputInterface $input)
+    {
+        /** @var Configuration $conf */
+        $conf = $this->getContainer()->get(Configuration::class);
+
+        if (Service::getParameter(self::NO_CONF))
+        {
+            Service::setWorkingDirectory(getcwd());
+            $conf->parse();
+        }
+        else
+        {
+            $confFilePath = $input->getOption('conf');
+            $siteRoot = fs::getFolderPath(realpath($confFilePath));
+            Service::setWorkingDirectory($siteRoot);
+
+            $configFile = new File($confFilePath);
+
+            $conf->parse($configFile);
+        }
+    }
+
+    /**
+     * Set a parameter to the Service singleton.
+     *
+     * @param InputInterface $input
+     * @param string         $param
+     */
+    private function setServiceParameter(InputInterface &$input, $param)
+    {
+        Service::setParameter($param, $input->getOption($param));
     }
 }
