+++ src/allejo/stakx/RuntimeStatus.php
@@ -1,0 +1,10 @@
+<?php
+
+
+namespace allejo\stakx;
+
+
+class RuntimeStatus
+{
+
+}
